prvi

+++ src/main/java/rs/raf/demo/controllers/StudentRestController.java
@@ -19,16 +19,15 @@ public class StudentRestController {
 
     private final StudentService studentService;
 
-
     public StudentRestController(StudentService studentService) {
         this.studentService = studentService;
     }
 
     @GetMapping(value = "/all",
-            produces = MediaType.APPLICATION_JSON_VALUE )
-    public List<Student> getAllStudents() {
+            produces = MediaType.APPLICATION_JSON_VALUE)
+    public List<Student> getAllStudents(){
         return studentService.findAll();
-    }
+    };
 
     @GetMapping(produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<?> getStudentById(@RequestParam("studentId") Long id){
@@ -43,19 +42,38 @@ public class StudentRestController {
     @PostMapping(consumes = MediaType.APPLICATION_JSON_VALUE,
             produces = MediaType.APPLICATION_JSON_VALUE)
     public Student createStudent(@RequestBody Student student){
-         return studentService.save(student);
+        return studentService.save(student);
     }
 
     @PutMapping(consumes = MediaType.APPLICATION_JSON_VALUE,
             produces = MediaType.APPLICATION_JSON_VALUE)
     public Student updateStudent(@RequestBody Student student){
+        System.out.println("updating");
         return studentService.save(student);
     }
 
     @DeleteMapping(value = "/{id}")
-    public ResponseEntity<?> deleteStudent(@PathVariable Long id){
-        studentService.deleteById(id);
-        return ResponseEntity.noContent().build();
+    public ResponseEntity<?> deleteStudent(@PathVariable Long id) {
+        Optional<Student> optionalStudent = studentService.findById(id);
+        if(optionalStudent.isPresent()) {
+            Student student = optionalStudent.get();
+
+            for (int i = 0; i < student.getCourses().size(); i++) {
+                student.getCourses().get(i).removeStudent(student);
+            }
+            studentService.deleteById(id);
+
+            return ResponseEntity.ok().build();
+        }
+        return ResponseEntity.notFound().build();
+    }
+
+    @GetMapping(value = "/search",
+            produces = MediaType.APPLICATION_JSON_VALUE)
+    public List<Student> searchStudent(@RequestParam String firstName,
+                                       @RequestParam String lastName){
+
+        return studentService.findByFirstNameAndLastName(firstName, lastName);
     }
 
 }
